codexchat_back/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
argon2-cffi==23.1.0
redis==6.0.0
